functions/json_to_sheets/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Exports Postgres DB to JSON"
authors = ["Nat Hillard <nathaniel.k.hillard@eop.omb.gov>"]

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.12.8"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.2"
pandas = "^1.2.1"


[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pylint = "^2.6.0"
autopep8 = "^1.5.4"
pydantic = {extras = ["dotenv"], version = "^1.7.3"}
importlib-resources = "^5.1.0"
pytest-coverage = "^0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
